Flytt kode ut i egen fil + flytt readme 1 mappe ut
